게시판 기능 수정

+++ src/main/java/com/ssafy/buyhouse/domain/board/service/BoardService.java
@@ -6,6 +6,7 @@ import com.ssafy.buyhouse.domain.board.dto.request.PostRequestDto;
 import com.ssafy.buyhouse.domain.board.dto.response.PostDetailResponseDto;
 import com.ssafy.buyhouse.domain.board.dto.response.PostResponseDto;
 import com.ssafy.buyhouse.domain.board.repository.BoardRepository;
+import com.ssafy.buyhouse.domain.member.domain.Member;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -32,8 +33,11 @@ public class BoardService {
         // board를 response객체로 변환
         return boards.map(board ->
                 PostResponseDto.builder()
+                        .id(board.getId())
                         .title(board.getTitle())
-                        .name(board.getMember().getName()) // 맴버넣으면 board.getMember().getName()
+                        .name(board.getMember().getName())
+                        .createdDate(board.getCreatedAt())
+                        .modifiedAt(board.getModifiedAt())// 맴버넣으면 board.getMember().getName()
                         .build());
     }
 
@@ -50,7 +54,7 @@ public class BoardService {
     }
 
     // 단일 게시물 상세 보기
-    public PostDetailResponseDto findById(long id) {
+    public PostDetailResponseDto findById(Long id) {
         Board board = boardRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("존재하지 않는 게시물입니다. id=" + id));
 
@@ -59,9 +63,9 @@ public class BoardService {
 
     // 게시물 작성
     @Transactional
-    public String save(PostRequestDto postRequestDto) { // Member memeber
+    public String save(PostRequestDto postRequestDto, Member member) { // Member memeber
         Board board = postRequestDto.toEntity(postRequestDto);
-        // board.setMember(member);
+        board.setMember(member);
         boardRepository.save(board);
 
         return "게시물이 저장되었습니다.";
@@ -69,9 +73,11 @@ public class BoardService {
 
     // 게시물 수정
     @Transactional
-    public String update(Long id, PostRequestDto postRequestDto) {
+    public String update(Long id, PostRequestDto postRequestDto, Member member) throws IllegalAccessException {
         Board board = boardRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("존재하지 않는 게시물입니다. id=" + id));
+
+        if(!board.getMember().getId().equals(member.getId())) throw new IllegalAccessException("해당 사용자의 게시물이 아닙니다.");
 
         board.setTitle(postRequestDto.getTitle());
         board.setContent(postRequestDto.getContent());
@@ -81,9 +87,11 @@ public class BoardService {
 
     // 게시물 삭제
     @Transactional
-    public String delete(Long id) {
+    public String delete(Long id, Member member) throws IllegalAccessException {
         Board board = boardRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("존재하지 않는 게시물입니다. id=" + id));
+
+        if(!board.getMember().getId().equals(member.getId())) throw new IllegalAccessException("해당 사용자의 게시물이 아닙니다.");
 
         boardRepository.delete(board);
